WIP add database function calls
